Replace derive_builder dependency with manual builder implementation
Replace the derive by exactly the same API, except the derived impl
block is merged with the hand-written impl block for the builder type.

Remove the docstrings for the private fields of Sqids that were
previously used to generate docs for the builder methods, so there is
only one place to update when improving the docs for these things. I'm
happy to revert this bit if you prefer having the internal docstrings.